# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CaliCoach API - calisthenics motion capture coaching app
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ExerciseDifficulty =
  (typeof ExerciseDifficulty)[keyof typeof ExerciseDifficulty];

export const ExerciseDifficulty = {
  beginner: "beginner",
  intermediate: "intermediate",
  advanced: "advanced",
} as const;

export interface Exercise {
  id: number;
  name: string;
  description: string;
  muscleGroups: string[];
  keyJoints: string[];
  difficulty: ExerciseDifficulty;
  coachingCues: string[];
}

export interface Session {
  id: number;
  exerciseId: number;
  exerciseName: string;
  startedAt: string;
  completedAt: string | null;
  totalReps: number;
  avgFormScore: number | null;
  notes: string | null;
}

export interface Rep {
  id: number;
  sessionId: number;
  repNumber: number;
  formScore: number;
  durationMs: number | null;
  feedbackGiven: string | null;
  loggedAt: string;
}

export type SessionWithReps = Session & {
  reps: Rep[];
};

export interface CreateSessionBody {
  exerciseId: number;
  notes?: string | null;
}

export interface UpdateSessionBody {
  completedAt?: string;
  totalReps?: number;
  avgFormScore?: number;
  notes?: string | null;
}

export interface CreateRepBody {
  repNumber: number;
  formScore: number;
  durationMs?: number | null;
  feedbackGiven?: string | null;
}

export interface ProgressSummary {
  totalSessions: number;
  totalReps: number;
  avgFormScore: number | null;
  bestFormScore: number | null;
  currentStreak: number;
  improvementPercent: number | null;
}

export interface ExerciseProgress {
  exerciseId: number;
  exerciseName: string;
  totalSessions: number;
  totalReps: number;
  avgFormScore: number | null;
  bestFormScore: number | null;
  lastSessionAt: string | null;
}

export interface TimelinePoint {
  date: string;
  avgFormScore: number;
  totalReps: number;
  exerciseId: number | null;
  exerciseName: string | null;
}

export interface SessionSummary {
  id: number;
  exerciseId: number;
  exerciseName: string;
  startedAt: string;
  totalReps: number;
  avgFormScore: number | null;
  durationMinutes: number | null;
}

export type ListSessionsParams = {
  limit?: number;
  offset?: number;
};

export type GetProgressTimelineParams = {
  exerciseId?: number;
  days?: number;
};

export type GetRecentSessionsParams = {
  limit?: number;
};