# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CaliCoach API - calisthenics motion capture coaching app
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sessions/{sessionId}/reps:
    get:
      operationId: listReps
      tags: [reps]
      summary: List reps for a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rep"
    post:
      operationId: createRep
      tags: [reps]
      summary: Log a rep for a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRepBody"
      responses:
        "201":
          description: Created rep
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rep"
...
components:
  schemas:
...
    Rep:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        repNumber:
          type: integer
        formScore:
          type: number
        durationMs:
          type: integer
          nullable: true
        feedbackGiven:
          type: string
          nullable: true
        loggedAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - repNumber
        - formScore
        - durationMs
        - feedbackGiven
        - loggedAt
    CreateRepBody:
      type: object
      properties:
        repNumber:
          type: integer
        formScore:
          type: number
        durationMs:
          type: integer
          nullable: true
        feedbackGiven:
          type: string
          nullable: true
      required:
        - repNumber
        - formScore
...