# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CaliCoach API - calisthenics motion capture coaching app
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all exercises
 */
export const ListExercisesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  muscleGroups: zod.array(zod.string()),
  keyJoints: zod.array(zod.string()),
  difficulty: zod.enum(["beginner", "intermediate", "advanced"]),
  coachingCues: zod.array(zod.string()),
});
export const ListExercisesResponse = zod.array(ListExercisesResponseItem);

/**
 * @summary Get a single exercise
 */
export const GetExerciseParams = zod.object({
  id: zod.coerce.number(),
});

export const GetExerciseResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  muscleGroups: zod.array(zod.string()),
  keyJoints: zod.array(zod.string()),
  difficulty: zod.enum(["beginner", "intermediate", "advanced"]),
  coachingCues: zod.array(zod.string()),
});

/**
 * @summary List all workout sessions
 */
export const listSessionsQueryLimitDefault = 20;
export const listSessionsQueryOffsetDefault = 0;

export const ListSessionsQueryParams = zod.object({
  limit: zod.coerce.number().default(listSessionsQueryLimitDefault),
  offset: zod.coerce.number().default(listSessionsQueryOffsetDefault),
});

export const ListSessionsResponseItem = zod.object({
  id: zod.number(),
  exerciseId: zod.number(),
  exerciseName: zod.string(),
  startedAt: zod.coerce.date(),
  completedAt: zod.coerce.date().nullable(),
  totalReps: zod.number(),
  avgFormScore: zod.number().nullable(),
  notes: zod.string().nullable(),
});
export const ListSessionsResponse = zod.array(ListSessionsResponseItem);

/**
 * @summary Start a new workout session
 */
export const CreateSessionBody = zod.object({
  exerciseId: zod.number(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get a single session with reps
 */
export const GetSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSessionResponse = zod
  .object({
    id: zod.number(),
    exerciseId: zod.number(),
    exerciseName: zod.string(),
    startedAt: zod.coerce.date(),
    completedAt: zod.coerce.date().nullable(),
    totalReps: zod.number(),
    avgFormScore: zod.number().nullable(),
    notes: zod.string().nullable(),
  })
  .and(
    zod.object({
      reps: zod.array(
        zod.object({
          id: zod.number(),
          sessionId: zod.number(),
          repNumber: zod.number(),
          formScore: zod.number(),
          durationMs: zod.number().nullable(),
          feedbackGiven: zod.string().nullable(),
          loggedAt: zod.coerce.date(),
        }),
      ),
    }),
  );

/**
 * @summary Update (complete) a session
 */
export const UpdateSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSessionBody = zod.object({
  completedAt: zod.coerce.date().optional(),
  totalReps: zod.number().optional(),
  avgFormScore: zod.number().optional(),
  notes: zod.string().nullish(),
});

export const UpdateSessionResponse = zod.object({
  id: zod.number(),
  exerciseId: zod.number(),
  exerciseName: zod.string(),
  startedAt: zod.coerce.date(),
  completedAt: zod.coerce.date().nullable(),
  totalReps: zod.number(),
  avgFormScore: zod.number().nullable(),
  notes: zod.string().nullable(),
});

/**
 * @summary List reps for a session
 */
export const ListRepsParams = zod.object({
  sessionId: zod.coerce.number(),
});

export const ListRepsResponseItem = zod.object({
  id: zod.number(),
  sessionId: zod.number(),
  repNumber: zod.number(),
  formScore: zod.number(),
  durationMs: zod.number().nullable(),
  feedbackGiven: zod.string().nullable(),
  loggedAt: zod.coerce.date(),
});
export const ListRepsResponse = zod.array(ListRepsResponseItem);

/**
 * @summary Log a rep for a session
 */
export const CreateRepParams = zod.object({
  sessionId: zod.coerce.number(),
});

export const CreateRepBody = zod.object({
  repNumber: zod.number(),
  formScore: zod.number(),
  durationMs: zod.number().nullish(),
  feedbackGiven: zod.string().nullish(),
});

/**
 * @summary Get overall progress summary stats
 */
export const GetProgressSummaryResponse = zod.object({
  totalSessions: zod.number(),
  totalReps: zod.number(),
  avgFormScore: zod.number().nullable(),
  bestFormScore: zod.number().nullable(),
  currentStreak: zod.number(),
  improvementPercent: zod.number().nullable(),
});

/**
 * @summary Get progress grouped by exercise
 */
export const GetProgressByExerciseResponseItem = zod.object({
  exerciseId: zod.number(),
  exerciseName: zod.string(),
  totalSessions: zod.number(),
  totalReps: zod.number(),
  avgFormScore: zod.number().nullable(),
  bestFormScore: zod.number().nullable(),
  lastSessionAt: zod.coerce.date().nullable(),
});
export const GetProgressByExerciseResponse = zod.array(
  GetProgressByExerciseResponseItem,
);

/**
 * @summary Form score timeline for charts
 */
export const getProgressTimelineQueryDaysDefault = 30;

export const GetProgressTimelineQueryParams = zod.object({
  exerciseId: zod.coerce.number().optional(),
  days: zod.coerce.number().default(getProgressTimelineQueryDaysDefault),
});

export const GetProgressTimelineResponseItem = zod.object({
  date: zod.coerce.date(),
  avgFormScore: zod.number(),
  totalReps: zod.number(),
  exerciseId: zod.number().nullable(),
  exerciseName: zod.string().nullable(),
});
export const GetProgressTimelineResponse = zod.array(
  GetProgressTimelineResponseItem,
);

/**
 * @summary Get recent sessions with summary data
 */
export const getRecentSessionsQueryLimitDefault = 5;

export const GetRecentSessionsQueryParams = zod.object({
  limit: zod.coerce.number().default(getRecentSessionsQueryLimitDefault),
});

export const GetRecentSessionsResponseItem = zod.object({
  id: zod.number(),
  exerciseId: zod.number(),
  exerciseName: zod.string(),
  startedAt: zod.coerce.date(),
  totalReps: zod.number(),
  avgFormScore: zod.number().nullable(),
  durationMinutes: zod.number().nullable(),
});
export const GetRecentSessionsResponse = zod.array(
  GetRecentSessionsResponseItem,
);